requirements.txt
Django==3.1.3 # Попробуйте угадать
django-taggit==1.3.0 # Теги для постов
Pillow==8.0.1 # Либа для картинок, без неё никак.
sorl-thumbnail==12.7.0 # Это либа для работы с картинками, но я ей не пользуюсь, но удалять не хочу, т.к. вдруг чё сломается
django-tinymce==3.2.0 # Нормальные формы в админку
twitter==1.18.0 # Пока не использую
django-comments-dab==2.5.1 # Комменты к постам
djangorestframework==3.12.2 # В основном он тупо для коментов
django-extensions==3.1.0 # Установил лишь для того, чтобы резетить бд, но вроде штука крутая, чекнуть подробнее
psycopg2-binary
django-mailer